requirements.txt
Django==2.2.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
